+++ src/main/java/com/authlete/common/types/ErrorCode.java
@@ -540,6 +540,18 @@ public enum ErrorCode
 
 
     /**
+     * The credential type is not supported.
+     *
+     * @see <a href="https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html"
+     *      >OpenID for Verifiable Credential Issuance</a>
+     *
+     * @since 3.81
+     * @since Authlete 3.0
+     */
+    unsupported_credential_type,
+
+
+    /**
      * The credential format is not supported.
      *
      * @see <a href="https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html"